Remove Tags
Previously, the specification required a "tags" file that contained a set of
arbitrary data.  This hasn't been used for anything particularly useful to
date, and it has been pointed out that allowing arbitrary files in metadata
can achieve the same ends.

Because of these observations, this change removes tags from the
specification.

Signed-off-by: Ben Hale <bhale@pivotal.io>
Signed-off-by: Ben Hale <bhale@vmware.com>