Removes BOM from buildpack API
The existing data format describes a Bill-of-Materials file that is never actually generated by the lifecycle. The platform API should specify the format and location of the BOM so that platforms can use this data in a consistent and predictable manner. Currently the BOM is embedded in the io.buildpacks.build.metadata label. PR #87 will add this label data forma to the platform API spec.

Signed-off-by: Emily Casey <ecasey@vmware.com>