Change defaults that depending on the working dir
This has been a frequent request of the lifecycle. Default file paths relative to the working dir are confusing because we provide no guidance on what the working directory should be when the platform executes the lifecycle. Making defaults relative to layers is better because, when phases are executed in separate containers these files should exist on a volume so they may be passed between phases. Right now if a user changes the layers dir they must also change the path of group.toml, plan.toml, etc so that they files are passed between phases. This is annoying. If the defaults are relative to <layers>, changing the layers dir "just works" without requiring a variety of other flags.

Signed-off-by: Emily Casey <ecasey@vmware.com>