Add buildpack API to io.buildpacks.build.metadata label
For newer platform API (0.10 and above) we need to know the buildpack API to know
if the process args are overridable or not (when there is more than one element in the command,
the process is definitely from a newer buildpack, but if there is only one element in the command,
the args could be overridable or not overridable depending on the buildpack API).

Having this information will allow platforms such as pack to display process information to end users.

The lifecycle is already adding processes[0].buildpackID to the label, this updates the spec to reflect
the current implementation.

Signed-off-by: Natalie Arellano <narellano@vmware.com>