Add `clear-env = true` note
The problem is described in https://github.com/buildpacks/docs/pull/841. While looking at the docs I was trying to determine if the platform env dir is provided when using `clear-env = true` and reading this paragraph in isolation it appeared that it was not.

It's now clear that "in the environment of `/bin/detect` or `/bin/build`" is referring to the specific "the environment" of the process and not the general "the environment" of the buildpack. However at the time, I somehow managed to mis-read this.

The sentence I'm adding, is redundant based on the paragraph above it, but I feel this helps clarify what's going on to a casual reader.

Signed-off-by: Schneems <richard.schneeman+foo@gmail.com>